education-assistant/requirements.txt
wikipedia
faiss-cpu
sentence-transformers
requests
groq
python-dotenv
